refactor: move search queries from `feed` schema to `search` schema
