+++ src/schema/search.ts
@@ -10,12 +10,27 @@ import {
   SearchSession,
 } from '../integrations';
 import { ValidationError } from 'apollo-server-errors';
-import { GQLEmptyResponse } from './common';
+import {
+  GQLEmptyResponse,
+  getSearchQuery,
+  offsetPageGenerator,
+} from './common';
 import { Connection as ConnectionRelay } from 'graphql-relay/connection/connection';
 import graphorm from '../graphorm';
 import { ConnectionArguments } from 'graphql-relay/index';
+import { FeedArgs, feedResolver } from '../common';
+import { GQLPost } from './posts';
 
 type GQLSearchSession = Pick<SearchSession, 'id' | 'prompt' | 'createdAt'>;
+
+interface GQLSearchPostSuggestion {
+  title: string;
+}
+
+export interface GQLSearchPostSuggestionsResults {
+  query: string;
+  hits: GQLSearchPostSuggestion[];
+}
 
 export const typeDefs = /* GraphQL */ `
   type SearchSession {
@@ -67,6 +82,15 @@ export const typeDefs = /* GraphQL */ `
     chunks: [SearchChunk]
   }
 
+  type SearchPostSuggestion {
+    title: String!
+  }
+
+  type SearchPostSuggestionsResults {
+    query: String!
+    hits: [SearchPostSuggestion!]!
+  }
+
   extend type Query {
     """
     Get user's search history
@@ -87,6 +111,46 @@ export const typeDefs = /* GraphQL */ `
     Get a search session by id
     """
     searchSession(id: String!): Search! @auth
+
+    """
+    Get suggestions for search post query
+    """
+    searchPostSuggestions(
+      """
+      The query to search for
+      """
+      query: String!
+    ): SearchPostSuggestionsResults!
+
+    """
+    Get a posts feed of a search query
+    """
+    searchPosts(
+      """
+      The query to search for
+      """
+      query: String!
+
+      """
+      Time the pagination started to ignore new items
+      """
+      now: DateTime
+
+      """
+      Paginate after opaque cursor
+      """
+      after: String
+
+      """
+      Paginate first
+      """
+      first: Int
+
+      """
+      Array of supported post types
+      """
+      supportedTypes: [String!]
+    ): PostConnection!
   }
 
   extend type Mutation {
@@ -96,6 +160,22 @@ export const typeDefs = /* GraphQL */ `
     searchResultFeedback(chunkId: String!, value: Int!): EmptyResponse! @auth
   }
 `;
+
+const searchResolver = feedResolver(
+  (ctx, { query }: FeedArgs & { query: string }, builder, alias) =>
+    builder
+      .andWhere(`${alias}.tsv @@ (${getSearchQuery(':query')})`, {
+        query,
+      })
+      .orderBy('views', 'DESC'),
+  offsetPageGenerator(30, 50),
+  (ctx, args, page, builder) => builder.limit(page.limit).offset(page.offset),
+  {
+    removeHiddenPosts: true,
+    removeBannedPosts: false,
+    allowPrivateSources: false,
+  },
+);
 
 export const resolvers: IResolvers<unknown, Context> = traceResolvers({
   Query: {
@@ -115,6 +195,42 @@ export const resolvers: IResolvers<unknown, Context> = traceResolvers({
     },
     searchSession: async (_, { id }: { id: string }, ctx): Promise<Search> =>
       getSession(ctx.userId, id),
+    searchPostSuggestions: async (
+      source,
+      { query }: { query: string },
+      ctx,
+    ): Promise<GQLSearchPostSuggestionsResults> => {
+      const hits: { title: string }[] = await ctx.con.query(
+        `
+            WITH search AS (${getSearchQuery('$1')})
+            select ts_headline(process_text(title), search.query,
+                               'StartSel = <strong>, StopSel = </strong>') as title
+            from post
+                   inner join search on true
+                   inner join source on source.id = post."sourceId"
+            where tsv @@ search.query and source.private = false
+            order by views desc
+              limit 5;
+          `,
+        [query],
+      );
+      return {
+        query,
+        hits,
+      };
+    },
+    searchPosts: async (
+      source,
+      args: FeedArgs & { query: string },
+      ctx,
+      info,
+    ): Promise<ConnectionRelay<GQLPost> & { query: string }> => {
+      const res = await searchResolver(source, args, ctx, info);
+      return {
+        ...res,
+        query: args.query,
+      };
+    },
   },
   Mutation: {
     searchResultFeedback: async (
